Fixed unhandled Not_found when probing terminfo
If `stty` returns a non-zero exit code, the code opts out of using `stty` to determine the terminal width by throwing a Not_found exception. However, this exception is not accounted for the next link of the try-with chain and thus terminates the search prematurely, crashing ocp-build, without checking $COLUMNS or falling back to the default.